+++ packages/twenty-ui/src/utilities/index.ts
@@ -13,7 +13,6 @@ export {
   stringToThemeColor,
   stringToThemeColorP3String,
 } from './color/utils/stringToThemeColorP3String';
-export { themeColorSchema } from './color/utils/themeColorSchema';
 export { getOsControlSymbol } from './device/getOsControlSymbol';
 export { getOsShortcutSeparator } from './device/getOsShortcutSeparator';
 export { getUserDevice } from './device/getUserDevice';